requirements.txt
kalshi-python>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
ruff>=0.1.0
mypy>=1.0.0
